Initial working merge of Snes9xgit + Snes9x-rr + Gtk
Snes9x-rr is a fork of Snes9x that adds support for scripting using Lua.
Lua scripting isn't just about automating the emulator. Lua scripts can
also export emulator internals. There are, however, a pair of problems
with RR as of this writing.

1) Most of the backend work of RR is cross-platform. However, frontend
   work for RR is focused on the Windows-only port.

2) RR has not been updated in 5 years, and is thus out of sync with
   upstream S9x.

My primary interest with this repository is to have an up-to-date S9x
Gtk port that has Lua scripting, so that I can have a working auto
tracker.

Most of the porting work was done here:
https://github.com/Nidhoegger/snes9x-rr . I took the scripting-related
code from that RR, and merged in into the latest S9x. Then I made it
compile (the tough part).

Bear in mind that this patch technically works, but it isn't pretty.
I'll make it look nice later on. :)