I am not interested in messages related to state load/save.
There are two kinds of messages that are being reported when a state is
saved or loaded: Terminal and emulator info (in emulator popups).

I don't see the value in reporting state changes to the terminal. I
don't want to suppress s9x messages entirely, but I don't want to see a
plethora of state change messages.

The in-emulator popups could be useful, but I, at this time, don't want
them. I might re-enable them in the future if a configuration option is
set to show them (and show loaded/saved slot <N> instead of a filename).
For the time being, this is acceptable.